// frontEnd/src/app/models/member.ts
export class Member {
    _id: string;
    firstName: string; 
    lastName: string;
    gender: string;    
    birth: string;    
    email: string;
    phone: string;
    cellphone: string; 
    address: string;
    postalCode: string;
    job: string;
    workplace: string;
    position: string; 
    serviceDept: string;
    serviceGroup: string;
    comment: string;
    createdDate: string;
    modifiedDate: string;
}